Update openchain-standards-model-provisions.0.5.md
Updated the key model provisions document based on the outcome of the Legal Work Group call of 2023-07-28.

Basically: Andrew's additions mark the close of the document excepting final review, and the risk grid format items move to the risk grid. The risk grid will be the next subject for our review once core document is published.