Updated as per https://github.com/OpenChain-Project/Reference-Material/issues/83
The alternation is:
When an item with open source software is delivered to another party information related to all the included open source must be provided.
changed to:
In general, when an item that includes open source software is delivered to another company or legal entity, information related to all the included open source should be provided. It is good practice to deliver this type of information even when distributing internally to ensure accurate record-keeping.